Chatbot-Task1/requirements.txt
fastapi[all]
sqlalchemy # The ORM for handling database operations.
psycopg2-binary # The PostgreSQL database adapter.
alembic # The database migration tool.